Handle deletion of errored instances
When a new instance is created and the state is in error,
the event "compute.instance.create.end" is never sent.

However, when we delete an errored instance the deletion event is
sent, but the entity doesn't exists in te database.

This fix avoid false positives in the critical queue.

Change-Id: I70a4ae92bda37909cfd6021e0311a69b7ad189bc